[MMG_CRPG] [Client + WebAPI] 유저 보유 캐릭터 장비 #4  아이템 장착 창에서 보유 아이템 보여주기

+++ WebServer/Domain/Entities/StatType.cs
@@ -8,7 +8,7 @@ namespace Domain.Entities
 {
     public sealed class StatType
     {
-        public short Id { get; private set; }          // PostgreSQL smallint ↔ C# short
+        public int Id { get; private set; }          // PostgreSQL smallint ↔ C# short
         public string Code { get; private set; } = ""; // UNIQUE
         public string Name { get; private set; } = "";
         public bool IsPercent { get; private set; }
